Safe_monitor/requirements.txt
mss>=9.0.0
easyocr>=1.6.0
opencv-python>=4.8.0
numpy>=1.24.3
Pillow>=10.0.0
twilio>=8.10.0
